bot/requirements.txt
requests==2.25.1
mysql.connector==2.2.9